Changed MotorDriver library to use a 4 byte C union to send over I2C as opposed to a string rep of float.

+++ MotorDriver/MotorDriver.h
@@ -34,7 +34,7 @@ public:
 private:
     long cmToSteps(float cm);
     void writeToWire(int code, char *str);
-    void writeToWire(int code, long val1, long val2);
+    void writeToWire(int code, float val1, float val2);
     float _leftSpeed;
     float _rightSpeed;
     int _address;
